pyproject.toml
[project]
name = "project"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "asm-differ",
    "ninja-syntax>=1.7.2",
    "splat64[mips]>=0.40.0",
]

[tool.uv.sources]
asm-differ = { git = "https://github.com/simonlindholm/asm-differ" }
